send- mainchain functionality complete

+++ src/XelsDesktopWalletApp/Models/SendConfirmation.cs
@@ -9,16 +9,16 @@ namespace XelsDesktopWalletApp.Models
     public class SendConfirmation
     {
         public TransactionBuilding Transaction { get; set; }
-        public Decimal TransactionFee { get; set; }
+        public double TransactionFee { get; set; }
         public string Cointype { get; set; }
     }
 
     public class SendConfirmationSC
     {
         public TransactionBuildingSidechain Transaction { get; set; }
-        public decimal TransactionFee { get; set; }
+        public double TransactionFee { get; set; }
         //public bool sidechainEnabled { get; set; }
-        public decimal OpReturnAmount { get; set; }
+        public double OpReturnAmount { get; set; }
         //public bool hasOpReturn { get; set; }
         public string cointype { get; set; }
     }
