+++ src/XelsDesktopWalletApp/Views/SendConfirmationMainChain.xaml.cs
@@ -30,7 +30,7 @@ namespace XelsDesktopWalletApp.Views
 
         private void bindData(SendConfirmation data)
         {
-            decimal amountSent = data.Transaction.FeeAmount - data.TransactionFee;
+            double amountSent = data.Transaction.FeeAmount - data.TransactionFee;
             this.AmountSent.Content = amountSent;
             this.AmountSentType.Content = data.Cointype;
